Guess I'm finally cleaning these up...
This is the first portion of the config files that includes the most
important ones. Others will hopefully follow soon.

Also, the `setup.nu` script might not even be working correctly because
I was clever enough to write it once I set up all links manually, so I
haven't run it at all. 🙃