Replace pyrefly with ty
Pyrefly unfortunately ignores all hidden directories by default, so it's
virtually impossible to make it type-check Python files that are present
in non-Python projects but nested in a hidden directory (like this repo).
I don't want to have to turn all such directories into Python projects
just for the sake of better code editor experience.

ty is still very much experimental, but that should not be a problem for
type checking and providing basic Language Server capabilities in simple
scripts. It also belongs to the same stack as uv and ruff, which I use,
so it's easier to figure it out configuration-wise.