Update `font` commands
I'm going to manage the installed Python versions with uv. However, now
I have Python installed globally via homebrew because fontforge depends
on it. I'd like to unlink it, but the font patcher imports it, so doing
that would break it.

This patch updates the `font patch` command so that it only runs Python
scripts that depend on fontforge via the fontforge command itself. This
enables the scripts to import that module without polluting my global
environment.

Additionally, this adds a new `font download` command that automatically
downloads preconfigured variants of the MonoLisa font because doing that
manually via UI for the nth time is annoying. (And I always forget which
font features need to be configured for the variant with preconfigured
ligatures.)